Migrate to Docker Compose version 3.4 and use Multi-Stage builds
Multi-Stage builds reduce our image size and allow us to deploy images that do not contain unnecessary build tools.

For local development the image used is `build` from the Dockerfile and no source is included in the image.  All files and packages are volume mounted at run time.
For testing in circleci, the yarn install step occurs within the build step and the source + packages exist in the image.
For deployment, the base image (with no build tools) is used and the source + installed packages are copied from the prod image.