Upgrade ssh2-sftp-client version to 2.5.0 for new SSH Key format support
API changes to sftp.get options require minor alterations to retrieve data with correct encoding options.  This also returns a String/Buffer instead of a Stream if `undefined` is passed as the second argument.

This version includes a bump to ssh2 version 0.8.2 which in turn includes ssh2-streams version 0.4.2 which has a rewrite of Private Key parsing logic.