StorageCredential Migration should await the saving of the records
This migration has only been applied to the DEV environment so far.  I used that environment to debug the issue and reverted the previous migration.  When this is deployed, the new version of this migration will be applied to DEV as if it were being run for the first time.