Ledger padding should be increased to support larger vaults
There are currently no users of the engine historical retrievals (Transmission has not been updated for this, yet).  The harm in changing the padding length is very minimal as it only impacts a few testing vaults and there is no current ShipChain use of this functionality.  All new vaults will not have any migration need.